KAIZEN-0: Forenkle og Fix Textarea
Fjernet en bug som viste seg når det ble mye input
