+++ src/felleskomponenter/textarea/Textarea.less
@@ -1,0 +1,9 @@
+.textarea-custom {
+    overflow-y: auto;
+}
+
+.textarea-custom-teller {
+    left: auto;
+    right: 0;
+    bottom: -24px;
+}
