+++ src/view/dialog/Dialog.less
@@ -32,7 +32,6 @@
     padding: 1rem;
     align-items: flex-end;
     .autosizing-textarea {
-        flex: 1;
         max-height: 7rem;
         @media (min-width: 800px) {
             max-height: 15rem;
@@ -52,9 +51,6 @@
     }
 }
 
-.meldingsfelt {
-    width: 100%;
-}
 .checkbox-block {
     display: flex;
     padding-left: 1rem;
