+++ src/view/dialog/NyDialog.less
@@ -39,7 +39,6 @@
         padding: 1rem 1rem 1rem 0;
         align-items: flex-end;
         .autosizing-textarea {
-            flex: 1;
             max-height: 7rem;
             @media (min-width: 800px) {
                 max-height: 15rem;
