--- src/view/henvendelse/HenvendelseTextarea.less
@@ -1,26 +1,0 @@
-.autosizing-textarea {
-    position: relative;
-    &__mirror {
-        position: relative;
-        left: -99999px;
-        white-space: pre-wrap;
-        word-break: break-word;
-        border: 1px solid transparent;
-        line-height: 1.375rem;
-        min-height: 4rem;
-        padding: 1rem 0.5rem 1.5rem;
-    }
-
-    &__textarea {
-        position: absolute;
-        top: 0;
-        height: 100%;
-        overflow-y: auto;
-    }
-
-    + .textarea--medMeta__teller {
-        left: initial;
-        right: 0;
-        bottom: -24px;
-    }
-}
